+++ src/GafferModule/ValuePlugBinding.cpp
@@ -160,6 +160,7 @@ void GafferModule::bindValuePlug()
 		.value( "Standard", ValuePlug::CachePolicy::Standard )
 		.value( "TaskCollaboration", ValuePlug::CachePolicy::TaskCollaboration )
 		.value( "TaskIsolation", ValuePlug::CachePolicy::TaskIsolation )
+		.value( "Default", ValuePlug::CachePolicy::Default )
 		.value( "Legacy", ValuePlug::CachePolicy::Legacy )
 	;