+++ src/GafferScene/SceneReader.cpp
@@ -135,9 +135,9 @@ ValuePlug::CachePolicy cachePolicyFromEnv( const char *name )
 		{
 			return ValuePlug::CachePolicy::TaskIsolation;
 		}
-		else if( !strcmp( cp, "Legacy" ) )
+		else if( !strcmp( cp, "Legacy" ) || !strcmp( cp, "Default" ) )
 		{
-			return ValuePlug::CachePolicy::Legacy;
+			return ValuePlug::CachePolicy::Default;
 		}
 		else
 		{
@@ -148,7 +148,7 @@ ValuePlug::CachePolicy cachePolicyFromEnv( const char *name )
 		}
 	}
 
-	return ValuePlug::CachePolicy::Legacy;
+	return ValuePlug::CachePolicy::Default;
 }
 
 const ValuePlug::CachePolicy g_objectCachePolicy = cachePolicyFromEnv( "GAFFERSCENE_SCENEREADER_OBJECT_CACHEPOLICY" );