+++ Changes.md
@@ -27,6 +27,7 @@ API
 ---
 
 - Process : Added `acquireCollaborativeResult()` method, providing an improved mechanism for multiple threads to collaborate on TBB tasks spawned by a single process they all depend on.
+- ValuePlug : Added `Default` CachePolicy and deprecated `Standard`, `TaskIsolation` and `Legacy` policies.
 
 1.3.5.0 (relative to 1.3.4.0)
 =======